@@ -33,18 +33,18 @@
         <unidal.plexus.version>2.5.8</unidal.plexus.version>
         <commons-pool.version>2.3</commons-pool.version>
         <commons-lang3.version>3.4</commons-lang3.version>
-        <commons-httpclient.version>4.5.2</commons-httpclient.version>
-        <spring-boot.version>2.6.15</spring-boot.version>
-        <spring-boot-test.version>2.6.15</spring-boot-test.version>
-        <spring.version>5.3.18</spring.version>
-        <spring-core.version>5.3.18</spring-core.version>
-        <spring-boot-maven-plugin.version>2.6.15</spring-boot-maven-plugin.version>
+        <commons-httpclient.version>5.3.1</commons-httpclient.version>
+        <spring-boot.version>3.2.5</spring-boot.version>
+        <spring-boot-test.version>3.2.5</spring-boot-test.version>
+        <spring.version>6.1.6</spring.version>
+        <spring-core.version>6.1.6</spring-core.version>
+        <spring-boot-maven-plugin.version>3.2.5</spring-boot-maven-plugin.version>
         <spring.velocity.version>1.4.3.RELEASE</spring.velocity.version>
         <netty3.version>3.10.5.Final</netty3.version>
         <ctrip-framework-foundation.version>1.8.26</ctrip-framework-foundation.version>
         <ctrip-super-rule.version>1.0.25</ctrip-super-rule.version>
         <ctrip-apollo.version>0.6.7</ctrip-apollo.version>
-        <cat.version>3.6.13</cat.version>
+        <cat.version>3.7.9-Java21</cat.version>
         <pyrolite.version>4.9</pyrolite.version>
         <plexus-utils.version>3.1.0</plexus-utils.version>
         <git-commit-id-plugin.version>2.2.3</git-commit-id-plugin.version>
@@ -69,7 +69,7 @@
         <jackson.version>2.10.3</jackson.version>
         <jedis.version>2.8.1.4-ctrip</jedis.version>
         <soa.version>2.24.55</soa.version>
-        <sso.version>0.1.3</sso.version>
+        <sso.version>1.0.1-Java21</sso.version>
         <vi.version>0.11.49</vi.version>
         <dal.version>2.4.50</dal.version>
         <ucs.version>0.7.0</ucs.version>
