@@ -3,8 +3,8 @@ package com.ctrip.xpipe.sso;
 
 import com.ctrip.xpipe.api.sso.LogoutHandler;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
