@@ -1,14 +1,13 @@
 package com.ctrip.framework.xpipe.redis.servlet.spring;
 
+import jakarta.servlet.DispatcherType;
+import jakarta.servlet.ServletException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 import org.springframework.core.env.Environment;
-
-import javax.servlet.DispatcherType;
-import javax.servlet.ServletException;
 
 @Configuration
 public class AutoConfiguration {
